linux/requirements.txt
pipewire_python
pyserial